feat(web): bot 扩展包列表改为市场同款卡片网格，二级页按市场详情页版式
- 列表复用 market-item-card，两列网格，卡片只有图标、名称、描述，操作在卡片右侧
- 二级页不再放在页头里，自带返回/操作行、大图标与标题、描述，与市场详情页一致
- 删除 package-row.vue
